backend mostly done

+++ StackOverflow/spring/src/main/java/edu/utcn/stackoverflow/controller/dto/UserOutDto.java
@@ -1,11 +1,10 @@
 package edu.utcn.stackoverflow.controller.dto;
 
 import lombok.Data;
-import lombok.ToString;
 
 @Data
 public class UserOutDto {
-    private Integer id;
+    private Long id;
     private String lastName;
     private String firstName;
     private String userName;
